ai-service/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
torch
torchvision
opencv-python
Pillow
timm
